Final touches to redux port
