+++ app/config.js
@@ -1,4 +1,6 @@
 const config = {
-'API_BASE': 'http://tadadev.klp.org.in/api/v1/',
-'DEV_API_BASE': 'http://localhost:8000/api/v1/'
+'PROD_SERVER_API_BASE': 'http://tadadev.klp.org.in/api/v1/',
+'DEV_API_BASE': 'http://localhost:8080/api/v1/'
 }
+
+module.exports = config
